+++ source/test.ts
@@ -8,10 +8,10 @@ import kava from 'kava'
 import { isIgnoredPath } from './index.js'
 
 // Tests
-kava.suite('ignorefs', function (suite) {
+kava.suite('ignorefs', function (suite, test) {
 	suite('common patterns', function (suite, test) {
 		// Prepare
-		const ignoreExpected: { [key: string]: boolean } = {
+		const ignoreExpected: { [basename: string]: boolean } = {
 			// Vim
 			'~': true,
 			'~something': true,
@@ -73,13 +73,13 @@ kava.suite('ignorefs', function (suite) {
 		}
 
 		// Tests
-		Object.keys(ignoreExpected).forEach(function (path) {
-			const resultExpected = ignoreExpected[path]
+		Object.keys(ignoreExpected).forEach(function (basename) {
+			const resultExpected = ignoreExpected[basename]
 			const testName = `${
 				resultExpected ? 'should' : 'should not'
-			} ignore ${path}`
+			} ignore ${basename}`
 			test(testName, function () {
-				const resultActual = isIgnoredPath(path)
+				const resultActual = isIgnoredPath({ basename })
 				equal(resultActual, resultExpected, 'ignored result was as expected')
 			})
 		})
@@ -87,24 +87,41 @@ kava.suite('ignorefs', function (suite) {
 
 	suite('ignore paths', function (suite, test) {
 		// Prepare
-		const ignoreExpected: { [key: string]: boolean } = {
+		const ignoreExpected: { [absolutePath: string]: boolean } = {
 			'/usr/awesome': false,
 			'/tmp/awesome': true,
 			'/usr/tmp/awesome': true,
 		}
 
 		// Tests
-		Object.keys(ignoreExpected).forEach(function (path) {
-			const resultExpected = ignoreExpected[path]
+		Object.keys(ignoreExpected).forEach(function (absolutePath) {
+			const resultExpected = ignoreExpected[absolutePath]
 			const testName = `${
 				resultExpected ? 'should' : 'should not'
-			} ignore ${path}`
+			} ignore ${absolutePath}`
 			test(testName, function () {
-				const resultActual = isIgnoredPath(path, {
-					ignorePaths: ['/tmp', '/usr/tmp'],
-				})
+				const resultActual = isIgnoredPath(
+					{ absolutePath },
+					{
+						ignorePaths: ['/tmp', '/usr/tmp'],
+					}
+				)
 				equal(resultActual, resultExpected, 'ignored result was as expected')
 			})
 		})
+	})
+
+	test('callback includes custom', function () {
+		const resultActual = isIgnoredPath(
+			Object.assign({ custom: true, basename: 'something' }),
+			{
+				ignoreCustomCallback(path: any) {
+					if (path.custom !== true || path.basename !== 'something')
+						throw new Error('callback did not include expected properties')
+					return true
+				},
+			}
+		)
+		equal(resultActual, true, 'ignored result was as expected')
 	})
 })